[#2483] user controller front end tests
register user test, move login tests to controller/test_user.py
