+++ ckan/tests/lib/test_base.py
@@ -1,89 +1,6 @@
 from nose import tools as nose_tools
 
-import ckan.tests.factories as factories
 import ckan.tests.helpers as helpers
-
-
-class TestLoginView(helpers.FunctionalTestBase):
-
-    def _maybe_follow(self, response, **kw):
-        """
-        Follow all redirects. If this response is not a redirect, do nothing.
-        Returns another response object.
-
-        (backported from WebTest 2.0.1)
-        """
-        remaining_redirects = 100  # infinite loops protection
-
-        while 300 <= response.status_int < 400 and remaining_redirects:
-            response = response.follow(**kw)
-            remaining_redirects -= 1
-
-        assert remaining_redirects > 0, "redirects chain looks infinite"
-        return response
-
-    def test_registered_user_login(self):
-        '''
-        Registered user can submit valid login details at /user/login and
-        be returned to appropriate place.
-        '''
-        app = helpers._get_test_app()
-
-        # make a user
-        user = factories.User()
-
-        # get the form
-        response = app.get('/user/login')
-        # ...it's the second one
-        login_form = response.forms[1]
-
-        # fill it in
-        login_form['login'] = user['name']
-        login_form['password'] = 'pass'
-
-        # submit it
-        submit_response = login_form.submit()
-        # let's go to the last redirect in the chain
-        final_response = self._maybe_follow(submit_response)
-
-        # the response is the user dashboard, right?
-        final_response.mustcontain('<a href="/dashboard">Dashboard</a>',
-                                   '<span class="username">{0}</span>'
-                                   .format(user['fullname']))
-        # and we're definitely not back on the login page.
-        final_response.mustcontain(no='<h1 class="page-heading">Login</h1>')
-
-    def test_registered_user_login_bad_password(self):
-        '''
-        Registered user is redirected to appropriate place if they submit
-        invalid login details at /user/login.
-        '''
-        app = helpers._get_test_app()
-
-        # make a user
-        user = factories.User()
-
-        # get the form
-        response = app.get('/user/login')
-        # ...it's the second one
-        login_form = response.forms[1]
-
-        # fill it in
-        login_form['login'] = user['name']
-        login_form['password'] = 'badpass'
-
-        # submit it
-        submit_response = login_form.submit()
-        # let's go to the last redirect in the chain
-        final_response = self._maybe_follow(submit_response)
-
-        # the response is the login page again
-        final_response.mustcontain('<h1 class="page-heading">Login</h1>',
-                                   'Login failed. Bad username or password.')
-        # and we're definitely not on the dashboard.
-        final_response.mustcontain(no='<a href="/dashboard">Dashboard</a>'),
-        final_response.mustcontain(no='<span class="username">{0}</span>'
-                                   .format(user['fullname']))
 
 
 class TestCORS(helpers.FunctionalTestBase):
